Update API product and seller product

+++ src/main/java/com/dinchan/model/Cart.java
@@ -23,7 +23,7 @@ public class Cart {
     @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL, orphanRemoval = true)
     private Set<CartItem> cartItems = new HashSet<>();
 
-    private Double totalSellingPrice;
+    private int totalSellingPrice;
     private int totalItem;
     private int totalMrpPrice;
     private int discount;
